Use conditional compilation to support Go without BoringSSL
The issue is that the import of crypto/boring requires the binary to be compiled with the boringcrypto tag which would force the cloudflare compilation to always use BoringSSL.
These changes:
- revert the cipher_suite.go changes and add conditional compilation.
- add xor_nonce_aead_boring.go
- add xor_nonce_aead_noboring.go
